Changed locations of maps to avoid compiling errors in SU2
